Improve the structure of the project
